Added prompts if the user isn't/is already in the birthday blacklist

+++ src/commands/blacklist/blacklist-remove-sub-command.ts
@@ -49,6 +49,16 @@ export class BlacklistRemoveSubCommand {
             return;
         }
 
+        let blacklist = await this.blacklistRepo.getBlacklist(msg.guild.id);
+
+        if (!blacklist.blacklist.map(entry => entry.UserDiscordId).includes(msg.author.id)) {
+            let embed = new MessageEmbed()
+                .setDescription('This user isn\'t in the blacklist!')
+                .setColor(Config.colors.error);
+            await channel.send(embed);
+            return;
+        }
+
         await this.blacklistRepo.removeBlacklist(msg.guild.id, target.id);
 
         let embed = new MessageEmbed()
